MatrixClient: require also the user_id when creating using a token
Although the user_id is not strictly needed for interacting with the server if
a token is already on hand, internally we sometimes need to know who we are.

Starting with this changeset, we require the user_id (not to be confused with
the username!) to be supplied when creatng a MatrixClient using an existing
token.

Note that a similar issue still exists regarding the HS (.hs member) -- when
registering or logging in with a password, the client stores the HS as returned
in the register/login response; however, no HS is set when creating with a
token. At the moment, the HS is not used anywhere, but this should probably be
addressed anyway -- see issue #38 for that...