matrix_client: Allow dynamic removal of listeners
Adding listeners is an action that cannot be reversed, while many users
need the ability to dynamically change whether they want to listen or not.
This patch assigns each listener with a unique id which allows users to
use it as a later reference for the same listener in case they want to
remove it.
This infrastructure can also be used in future commits to allow modifying
of listeners on-the-go.

Signed-off-by: Gal Pressman <galpressman@gmail.com>